working video simulation
